fix(shift): :poop: fix circular dependcies
the converter in the employee class caused a circular depency when loading an employee form firestore when setting it's employee value. Majorly refactored how the pay is caluclated and moved up the pay to the employee class and added the employee object (poorly💩) to the paystub class and the pay is calulated there. I also added a wage and pay properties to the employee class

probably🤷🏿‍♂️